refactor: 지하철 노선 생성 패키지 정리
