+++ src/main/java/subway/line/create/LineCreateRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.create;
 
 public class LineCreateRequest {
